`parseBlocksForUniquePublicKeys`: Move in `db/kv`.
