+++ validator/db/kv/import_test.go
@@ -168,6 +168,250 @@ func TestStore_ImportInterchangeData_OK(t *testing.T) {
 	}
 }
 
+func Test_parseUniqueSignedBlocksByPubKey(t *testing.T) {
+	numValidators := 4
+	publicKeys, err := valtest.CreateRandomPubKeys(numValidators)
+	require.NoError(t, err)
+	roots := valtest.CreateMockRoots(numValidators)
+	tests := []struct {
+		name    string
+		data    []*format.ProtectionData
+		want    map[[fieldparams.BLSPubkeyLength]byte][]*format.SignedBlock
+		wantErr bool
+	}{
+		{
+			name: "nil values are skipped",
+			data: []*format.ProtectionData{
+				{
+					Pubkey: fmt.Sprintf("%x", publicKeys[0]),
+					SignedBlocks: []*format.SignedBlock{
+						{
+							Slot:        "1",
+							SigningRoot: fmt.Sprintf("%x", roots[0]),
+						},
+						nil,
+					},
+				},
+				{
+					Pubkey: fmt.Sprintf("%x", publicKeys[0]),
+					SignedBlocks: []*format.SignedBlock{
+						{
+							Slot:        "3",
+							SigningRoot: fmt.Sprintf("%x", roots[2]),
+						},
+					},
+				},
+			},
+			want: map[[fieldparams.BLSPubkeyLength]byte][]*format.SignedBlock{
+				publicKeys[0]: {
+					{
+						Slot:        "1",
+						SigningRoot: fmt.Sprintf("%x", roots[0]),
+					},
+					{
+						Slot:        "3",
+						SigningRoot: fmt.Sprintf("%x", roots[2]),
+					},
+				},
+			},
+		},
+		{
+			name: "same blocks but different public keys are parsed correctly",
+			data: []*format.ProtectionData{
+				{
+					Pubkey: fmt.Sprintf("%x", publicKeys[0]),
+					SignedBlocks: []*format.SignedBlock{
+						{
+							Slot:        "1",
+							SigningRoot: fmt.Sprintf("%x", roots[0]),
+						},
+						{
+							Slot:        "2",
+							SigningRoot: fmt.Sprintf("%x", roots[1]),
+						},
+					},
+				},
+				{
+					Pubkey: fmt.Sprintf("%x", publicKeys[1]),
+					SignedBlocks: []*format.SignedBlock{
+						{
+							Slot:        "1",
+							SigningRoot: fmt.Sprintf("%x", roots[0]),
+						},
+						{
+							Slot:        "2",
+							SigningRoot: fmt.Sprintf("%x", roots[1]),
+						},
+					},
+				},
+			},
+			want: map[[fieldparams.BLSPubkeyLength]byte][]*format.SignedBlock{
+				publicKeys[0]: {
+					{
+						Slot:        "1",
+						SigningRoot: fmt.Sprintf("%x", roots[0]),
+					},
+					{
+						Slot:        "2",
+						SigningRoot: fmt.Sprintf("%x", roots[1]),
+					},
+				},
+				publicKeys[1]: {
+					{
+						Slot:        "1",
+						SigningRoot: fmt.Sprintf("%x", roots[0]),
+					},
+					{
+						Slot:        "2",
+						SigningRoot: fmt.Sprintf("%x", roots[1]),
+					},
+				},
+			},
+		},
+		{
+			name: "disjoint sets of signed blocks by the same public key are parsed correctly",
+			data: []*format.ProtectionData{
+				{
+					Pubkey: fmt.Sprintf("%x", publicKeys[0]),
+					SignedBlocks: []*format.SignedBlock{
+						{
+							Slot:        "1",
+							SigningRoot: fmt.Sprintf("%x", roots[0]),
+						},
+						{
+							Slot:        "2",
+							SigningRoot: fmt.Sprintf("%x", roots[1]),
+						},
+					},
+				},
+				{
+					Pubkey: fmt.Sprintf("%x", publicKeys[0]),
+					SignedBlocks: []*format.SignedBlock{
+						{
+							Slot:        "3",
+							SigningRoot: fmt.Sprintf("%x", roots[2]),
+						},
+					},
+				},
+			},
+			want: map[[fieldparams.BLSPubkeyLength]byte][]*format.SignedBlock{
+				publicKeys[0]: {
+					{
+						Slot:        "1",
+						SigningRoot: fmt.Sprintf("%x", roots[0]),
+					},
+					{
+						Slot:        "2",
+						SigningRoot: fmt.Sprintf("%x", roots[1]),
+					},
+					{
+						Slot:        "3",
+						SigningRoot: fmt.Sprintf("%x", roots[2]),
+					},
+				},
+			},
+		},
+		{
+			name: "full duplicate entries are uniquely parsed",
+			data: []*format.ProtectionData{
+				{
+					Pubkey: fmt.Sprintf("%x", publicKeys[0]),
+					SignedBlocks: []*format.SignedBlock{
+						{
+							Slot:        "1",
+							SigningRoot: fmt.Sprintf("%x", roots[0]),
+						},
+					},
+				},
+				{
+					Pubkey: fmt.Sprintf("%x", publicKeys[0]),
+					SignedBlocks: []*format.SignedBlock{
+						{
+							Slot:        "1",
+							SigningRoot: fmt.Sprintf("%x", roots[0]),
+						},
+					},
+				},
+			},
+			want: map[[fieldparams.BLSPubkeyLength]byte][]*format.SignedBlock{
+				publicKeys[0]: {
+					{
+						Slot:        "1",
+						SigningRoot: fmt.Sprintf("%x", roots[0]),
+					},
+					{
+						Slot:        "1",
+						SigningRoot: fmt.Sprintf("%x", roots[0]),
+					},
+				},
+			},
+		},
+		{
+			name: "intersecting duplicate public key entries are handled properly",
+			data: []*format.ProtectionData{
+				{
+					Pubkey: fmt.Sprintf("%x", publicKeys[0]),
+					SignedBlocks: []*format.SignedBlock{
+						{
+							Slot:        "1",
+							SigningRoot: fmt.Sprintf("%x", roots[0]),
+						},
+						{
+							Slot:        "2",
+							SigningRoot: fmt.Sprintf("%x", roots[1]),
+						},
+					},
+				},
+				{
+					Pubkey: fmt.Sprintf("%x", publicKeys[0]),
+					SignedBlocks: []*format.SignedBlock{
+						{
+							Slot:        "2",
+							SigningRoot: fmt.Sprintf("%x", roots[1]),
+						},
+						{
+							Slot:        "3",
+							SigningRoot: fmt.Sprintf("%x", roots[2]),
+						},
+					},
+				},
+			},
+			want: map[[fieldparams.BLSPubkeyLength]byte][]*format.SignedBlock{
+				publicKeys[0]: {
+					{
+						Slot:        "1",
+						SigningRoot: fmt.Sprintf("%x", roots[0]),
+					},
+					{
+						Slot:        "2",
+						SigningRoot: fmt.Sprintf("%x", roots[1]),
+					},
+					{
+						Slot:        "2",
+						SigningRoot: fmt.Sprintf("%x", roots[1]),
+					},
+					{
+						Slot:        "3",
+						SigningRoot: fmt.Sprintf("%x", roots[2]),
+					},
+				},
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := parseBlocksForUniquePublicKeys(tt.data)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("parseBlocksForUniquePublicKeys() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("parseBlocksForUniquePublicKeys() got = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
 func Test_parseUniqueSignedAttestationsByPubKey(t *testing.T) {
 	numValidators := 4
 	publicKeys, err := valtest.CreateRandomPubKeys(numValidators)
