+++ validator/helpers/BUILD.bazel
@@ -38,7 +38,6 @@ go_test(
         "//validator/db/common:go_default_library",
         "//validator/db/iface:go_default_library",
         "//validator/slashing-protection-history/format:go_default_library",
-        "//validator/testing:go_default_library",
         "@com_github_prometheus_client_golang//prometheus:go_default_library",
     ],
 )
